Minor fix ship

+++ BattleShip_4330153/src/Model/Ship.cs
@@ -3,7 +3,7 @@ using Microsoft.VisualBasic;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Data;
+//using System.Data;
 using System.Diagnostics;
 /// <summary>
 /// A Ship has all the details about itself. For example the shipname,
@@ -80,7 +80,7 @@ public class Ship
 		_tiles = new List<Tile>();
 
 		//gets the ship size from the enumarator
-		_sizeOfShip = _shipName;
+		_sizeOfShip = (int)_shipName;
 	}
 
 	/// <summary>
